requirements.txt
matplotlib
pandas
streamlit
flask
flask-cors
torch
torchvision
pillow
numpy
scikit-learn
